requirements.txt
pandas==1.5.3
streamlit==1.16.0
requests==2.28.1
joblib==1.3.2
millify==0.1.1
mlflow==2.6.0
boto3==1.28.28
snowflake-connector-python==2.8.3
geopy==2.3.0